int -> size_t in paper tests

+++ test/paper_tests.cpp
@@ -40,35 +40,35 @@ class MaxHeapStrategy : public Algorithm {
   Meta m("lcpcomp_strategy", "heap");
   return m; }
  using Algorithm::Algorithm;
- void create_factor(int pos,int src,int len);
- void factorize(text_t& text, const int t) {
+ void create_factor(size_t pos,size_t src,size_t len);
+ void factorize(text_t& text, const size_t t) {
   text.require(text_t::SA | text_t::ISA | text_t::LCP);
   auto& sa = text.require_sa();
   auto& isa = text.require_isa();
   auto lcpp = text.release_lcp()->relinquish();
   auto& lcp = *lcpp;
   ArrayMaxHeap<typename text_t::lcp_type::data_type> heap(lcp, lcp.size(), lcp.size());
-  for(int i = 1; i < lcp.size(); ++i)
+  for(size_t i = 1; i < lcp.size(); ++i)
    if(lcp[i] >= t) heap.insert(i);
   while(heap.size() > 0) {
-   int i = heap.top(), fpos = sa[i],
+   size_t i = heap.top(), fpos = sa[i],
        fsrc = sa[i-1], flen = heap.key(i);
    create_factor(fpos, fsrc, flen);
-   for(int k=0; k < flen; k++)
+   for(size_t k=0; k < flen; k++)
     heap.remove(isa[fpos + k]);
-   for(int k=0; k < flen && fpos > k; k++) {
-    int s = fpos - k - 1;
-    int j = isa[s];
+   for(size_t k=0; k < flen && fpos > k; k++) {
+    size_t s = fpos - k - 1;
+    size_t j = isa[s];
     if(heap.contains(j)) {
      if(s + lcp[j] > fpos) {
-      int l = fpos - s;
+      size_t l = fpos - s;
       if(l >= t)
        heap.decrease_key(j, l);
       else heap.remove(j);
 }}}}}};
 /////////////////////////////////
 template<class T>
-void MaxHeapStrategy<T>::create_factor(int pos, int src, int len){ /* [...] */ }
+void MaxHeapStrategy<T>::create_factor(size_t pos, size_t src, size_t len){ /* [...] */ }
 
 TEST(SEA17, MaxHeap) {
     auto text_ds = builder<TextDS<>>().instance("abc\0"_v);
@@ -76,21 +76,21 @@ TEST(SEA17, MaxHeap) {
     maxheap.factorize(text_ds, 1);
 }
 /////////////////////////////////
-void factorize(TextDS<>& T, SuffixTree& ST, std::function<void(int begin, int end, int ref)> output){
+void factorize(TextDS<>& T, SuffixTree& ST, std::function<void(size_t begin, size_t end, size_t ref)> output){
  typedef SuffixTree::node_type node_t;
  sdsl::int_vector<> R(ST.internal_nodes,0,bits_for(T.size() * bits_for(ST.cst.csa.sigma) / bits_for(T.size())));
- int pos = 0, z = 0;
+ size_t pos = 0, z = 0;
  while(pos < T.size() - 1) {
   node_t l = ST.select_leaf(ST.cst.csa.isa[pos]);
-  int leaflabel = pos;
+  size_t leaflabel = pos;
   if(ST.parent(l) == ST.root || R[ST.nid(ST.parent(l))] != 0) {
-   int parent_strdepth = ST.str_depth(ST.parent(l));
+   size_t parent_strdepth = ST.str_depth(ST.parent(l));
    output(pos + parent_strdepth, pos + parent_strdepth + 1, R[ST.nid(ST.parent(l))]);
    pos += parent_strdepth+1;
    ++z;
    continue;
   }
-  int d = 1;
+  size_t d = 1;
   node_t parent = ST.root;
   node_t node = ST.level_anc(l, d);
   while(R[ST.nid(node)] != 0) {
@@ -98,8 +98,8 @@ void factorize(TextDS<>& T, SuffixTree& ST, std::function<void(int begin, int en
    node = ST.level_anc(l, ++d);
   }
   pos += ST.str_depth(parent);
-  int begin = leaflabel + ST.str_depth(parent);
-  int end = leaflabel + ST.str_depth(node);
+  size_t begin = leaflabel + ST.str_depth(parent);
+  size_t end = leaflabel + ST.str_depth(node);
   output(begin, end, R[ST.nid(ST.parent(node))]);
   R[ST.nid(node)] = ++z;
   pos += end - begin;
@@ -109,8 +109,8 @@ void factorize(TextDS<>& T, SuffixTree& ST, std::function<void(int begin, int en
 TEST(SEA17, factorize) {
     auto T = "aaababaaabaababa\0"_v;
     auto text_ds = builder<TextDS<>>().instance(T);
-    std::vector<int> v;
-    auto f = [&](int begin, int end, int ref) {
+    std::vector<size_t> v;
+    auto f = [&](size_t begin, size_t end, size_t ref) {
         v.push_back(begin);
         v.push_back(end);
         v.push_back(ref);
@@ -124,7 +124,7 @@ TEST(SEA17, factorize) {
 
     factorize(text_ds, ST,f);
 
-    ASSERT_EQ(v, (std::vector<int> {
+    ASSERT_EQ(v, (std::vector<size_t> {
         0,   1,   0,
         2,   3,   1,
         3,   5,   0,
